10.2 add a nwe layer, the Bind class
Remove from the controller the responsibility to bind the model and its
view together with the handler.
This is now up to the Bind class.
Inside it we call the proxy factory, bind the model with the view and
call the handler.
Both the handler for message view and negotiations view share the same
pattern, "view.update(model)". This is the reason we could abstract it
even more.